refactor: extract getContentDisposition and FileResponse to own files
One export per file: move getContentDisposition to its own util and
FileResponse type to its own type file.

Made-with: Cursor
