+++ packages/twenty-server/src/engine/core-modules/file/types/file-response.type.ts
@@ -1,0 +1,5 @@
+import { type Readable } from 'stream';
+
+export type FileResponse =
+  | { type: 'redirect'; presignedUrl: string }
+  | { type: 'stream'; stream: Readable; mimeType: string };
